Major Blog Overhaul
I've had a lot of ideas lately about how the blog could function going forward. I clearly do not have enough time to be the only author. I plan to allow everyone to post with a peer review system. In order for others to post content though, I can't have the posts stored in fs as that would require redeployment for each post.

This new system relies on firebase storage to host author avatars, post content and cover images. Firebase Resize was set up to automatically produce a 400x800 and 1920x3840 copy of each image, one for the thumbnail and another for the larger hero image at the top. I'm still not sure how I'll manage images in the middle of the blog content but that will come later. Next.js was updated to pull all blog content from firestore and all blog data from the realtime database.

I was also able to implement easy/fast/bio reading in this PR! There will be a LOT more customization to come and even a text to speech option.